refactor(car): move delete write out of modal into section controls
DeleteModal now only renders. The mutation, the cached car-name read
used for the success toast, and the post-delete redirect move to a
co-located hook on SectionControls, its only caller, matching the
service-log and ownership delete modals.

canDelete now also disables the modal's confirm button, not just the
trigger. Previously a user demoted while the dialog was already open
could still submit the delete.
