+++ src/module/car/presentation/ui/modals/delete/delete.tsx
@@ -5,36 +5,34 @@ import { TextSeparator } from '@/ui/decorative/text-separator/text-separator';
 import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
 import { DialogModal } from '@/ui/dialog-modal/dialog-modal';
 
-import type { UseDeleteModalOptions } from './use-delete';
-import { useDeleteModal } from './use-delete';
-
 export const DELETE_MODAL_TEST_ID = 'delete-modal';
 
-type DeleteModalProps = UseDeleteModalOptions & {
+type DeleteModalProps = {
+  canDelete: boolean;
   ref?: RefObject<DialogModalRef | null>;
+  onCancel?: () => void;
+  onConfirm?: () => void;
+  onClose?: () => void;
 };
 
 export function DeleteModal({
-  carId,
+  canDelete,
   ref,
   onCancel,
   onConfirm,
+  onClose,
 }: DeleteModalProps) {
-  const {
-    handlers: { handleCancelButtonClick, handleDeleteButtonClick },
-  } = useDeleteModal({ carId, onCancel, onConfirm });
-
   return (
     <DialogModal ref={ref}>
-      <DialogModal.Root data-testid={DELETE_MODAL_TEST_ID}>
+      <DialogModal.Root data-testid={DELETE_MODAL_TEST_ID} onClose={onClose}>
         <DialogModal.Heading>Delete car</DialogModal.Heading>
         <TextSeparator className="my-4" />
         <p className="text-warning-500 dark:text-warning-300 my-4">
           Are you sure you want permanently delete this car?
         </p>
         <DialogModal.Controls>
-          <Button onClick={handleCancelButtonClick}>Cancel</Button>
-          <Button variant="error" onClick={handleDeleteButtonClick}>
+          <Button onClick={onCancel}>Cancel</Button>
+          <Button disabled={!canDelete} variant="error" onClick={onConfirm}>
             Delete
           </Button>
         </DialogModal.Controls>
